+++ Classes/Domain/Model/QuizSession.php
@@ -188,6 +188,15 @@ class QuizSession extends \TYPO3\CMS\Extbase\DomainObject\AbstractEntity
     }
 
     /**
+     * Deliver the question of the current step
+     * @return \Wacon\Simplequiz\Domain\Model\Question
+     */
+    public function getNextQuestion(): ?Question
+    {
+        return \array_key_exists($this->step-1, $this->questions) ? $this->questions[$this->step-1] : null;
+    }
+
+    /**
      * Set questions
      *
      * @param  array  $questions  Questions