+++ Classes/Domain/Repository/QuestionRepository.php
@@ -3,6 +3,8 @@
 declare(strict_types=1);
 
 namespace Wacon\Simplequiz\Domain\Repository;
+use TYPO3\CMS\Extbase\Persistence\Generic\QueryResult;
+use Wacon\Simplequiz\Domain\Model\Answer;
 
 /**
  * This file is part of the "Simplequiz" Extension for TYPO3 CMS.
@@ -16,4 +18,19 @@ namespace Wacon\Simplequiz\Domain\Repository;
 /**
  * The repository for Questions
  */
-class QuestionRepository extends \TYPO3\CMS\Extbase\Persistence\Repository {}
+class QuestionRepository extends \TYPO3\CMS\Extbase\Persistence\Repository {
+    /**
+     * Find one question record by answer
+     * @param \Wacon\Simplequiz\Domain\Model\Answer $answer
+     * @return \TYPO3\CMS\Extbase\Persistence\Generic\QueryResult
+     */
+    public function findOneByAnswer(Answer $answer): QueryResult
+    {
+        $query = $this->createQuery();
+        $query->matching(
+            $query->contains('answers', $answer)
+        );
+
+        return $query->execute();
+    }
+}